quota operation

+++ src/main/java/com/quiz/quotamanage/data/QuotaAccountDto.java
@@ -17,7 +17,7 @@ public class QuotaAccountDto {
     /**
      * 账户类型
      */
-    private Integer accountType;
+    private Byte accountType;
 
     /**
      * 账户额度
